+++ application/views/admin/kebutuhan/index.php
@@ -25,9 +25,11 @@
 
 					<!-- DataTables -->
 					<div class="card mb-3">
+						<?php if($this->session->user_logged["role_id"] == "1"){ ?>
 						<div class="card-header">
 							<a href="<?php echo site_url('dashboard/kebutuhan/add') ?>"><i class="fas fa-plus"></i> Add New</a>
 						</div>
+						<?php } ?>
 						<div class="card-body">
 
 							<div class="table-responsive">
@@ -40,7 +42,9 @@
 											<th>Jumlah</th>
 											<th>Satuan</th>
 											<th>Keterangan</th>
+						<?php if($this->session->user_logged["role_id"] == "1"){ ?>
 											<th>Aksi</th>
+						<?php } ?>
 										</tr>
 									</thead>
 									<tbody>
@@ -64,12 +68,14 @@
 											<td>
 												<?php echo $butuh->keterangan ?>
 											</td>
+						<?php if($this->session->user_logged["role_id"] == "1"){ ?>
 											<td width="250">
 												<a href="<?php echo site_url('dashboard/kebutuhan/edit/').$butuh->kebutuhan_id ?>"
 												class="btn btn-small"><i class="fas fa-edit"></i> Edit</a>
 												<a onclick="deleteConfirm('<?php echo $butuh->kebutuhan_id ?>')"
 												href="#!" class="btn btn-small text-danger"><i class="fas fa-trash"></i> Hapus</a>
 											</td>
+						<?php } ?>
 										</tr>
 										<?php endforeach; ?>
